Made the schedule for mobile view
