+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
 import HeroSection from "./components/herosection";
 import Navbar from "./components/navbar";
-import Schedule from "./components/schedule";
+import Schedule from "./pages/schedule";
 import ProjectsPage from "./pages/ProjectsPage";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
