+++ src/data/events.js
@@ -57,4 +57,15 @@
     }
   ];
 
-  export default events;
+const getEventTypeColor = (type) => {
+    switch (type) {
+      case 'ceremony': return 'from-purple-500 to-pink-500';
+      case 'workshop': return 'from-blue-500 to-purple-500';
+      case 'session': return 'from-green-500 to-teal-500';
+      case 'panel': return 'from-orange-500 to-red-500';
+      case 'sprint': return 'from-yellow-500 to-orange-500';
+      default: return 'from-purple-500 to-blue-500';
+    }
+  };
+
+  export { events, getEventTypeColor };
